+++ backupfiles/HooksBackup.ts
@@ -12,8 +12,20 @@ export let readyHooks = async () => {
 
   // setup all the hooks
 
+  // if(<boolean>game.settings.get(MODULE_NAME, "enabledShowDoorIcons")) {
+  //   //@ts-ignore
+  //   libWrapper.register(MODULE_NAME, 'ControlsLayer.prototype.drawDoors', ControlsLayerPrototypeDrawDoorsHandler, 'MIXED');
+  //   //@ts-ignore
+  //   libWrapper.register(MODULE_NAME, 'Wall.prototype._onModifyWall', WallPrototypeOnModifyWallHandler, 'MIXED');
+  //   //@ts-ignore
+  //   libWrapper.register(MODULE_NAME, 'WallsLayer.prototype.activate', WallsLayerPrototypeActivateHandler, 'MIXED');
+  // }
+
   Hooks.on('preUpdateWall', async (object, updateData, diff, userID) => {
 
+    // if(<boolean>game.settings.get(MODULE_NAME, "enableAmbientDoor")) {
+    //   AmbientDoors.preUpdateWallHandler(object, updateData, diff, userID);
+    // }
     // THIS IS ONLY A BUG FIXING FOR THE SOUND DISABLE FOR THE lib-wrapper override
     if(<boolean>game.settings.get(MODULE_NAME, "enableArmsReach")) {
       // if ambient door is present and active dont' do this
@@ -23,6 +35,56 @@ export let readyHooks = async () => {
     }
 
   });
+
+
+  // Hooks.on("renderWallConfig", (app, html, data) => {
+
+  //   if(<boolean>game.settings.get(MODULE_NAME, "enableAmbientDoor")) {
+  //     AmbientDoors.renderWallConfigHandler(app, html, data);
+  //   }
+
+  //   if(<boolean>game.settings.get(MODULE_NAME, "enableDesignerDoor")) {
+  //     DesignerDoors.renderWallConfigHandler(app, html, data);
+  //   }
+
+  //   if(<boolean>game.settings.get(MODULE_NAME, "enableWindowDoor")) {
+  //     WindowDoors.renderWallConfigHandler(app, html, data);
+  //   }
+
+  // });
+
+  // Hooks.on('renderFilePicker', (_app, html, _data) => {
+
+  //   if(<boolean>game.settings.get(MODULE_NAME, "enableSoundPreviewer")) {
+  //     SoundPreviewer.start(html);
+  //   }
+
+  // });
+
+  // Hooks.on('closeFilePicker', () => {
+
+  //   // Sound Preview
+
+  //   if(<boolean>game.settings.get(MODULE_NAME, "enableSoundPreviewer")) {
+  //     SoundPreviewer.stop();
+  //   }
+  // });
+
+  // Hooks.on("renderSettingsConfig", (app, html, user) => {
+
+  //   if(<boolean>game.settings.get(MODULE_NAME, "enableDesignerDoor")) {
+  //     DesignerDoors.renderSettingsConfigHandler(app, html, user);
+  //   }
+
+  // });
+
+  // Hooks.on('canvasInit', () => {
+
+  //   if(<boolean>game.settings.get(MODULE_NAME, "enableDesignerDoor")) {
+  //     DesignerDoors.canvasInitHandler();
+  //   }
+
+  // });
 
   // Management of the Stairways module
   if (game.modules.get("stairways")?.active){
@@ -71,7 +133,13 @@ export let readyHooks = async () => {
 
 export let setupHooks = () => {
 
+  // if(<boolean>game.settings.get(MODULE_NAME, "enableDesignerDoor")) {
+  //   //@ts-ignore
+  //   libWrapper.register(MODULE_NAME, 'DoorControl.prototype._getTexture', DoorControlPrototypeGetTextureHandler, 'OVERRIDE');
+  //   //@ts-ignore
+  //   // libWrapper.register(MODULE_NAME, 'DoorControl.prototype.draw', DoorControlPrototypeDrawHandler, 'MIXED');
 
+  // }
 }
 
 
@@ -83,13 +151,24 @@ export let initHooks = () => {
     Armsreach.init();
   }
 
+  // if(<boolean>game.settings.get(MODULE_NAME, "enableDesignerDoor")) {
+  //   DesignerDoors.init();
+  // }
+
+  //@ts-ignore
+  //libWrapper.register(MODULE_NAME, 'DoorControl.prototype._onMouseOver', DoorControlPrototypeOnMouseOverHandler, 'WRAPPER');
+
   if(<boolean>game.settings.get(MODULE_NAME, "enableArmsReach")) {
     //@ts-ignore
     libWrapper.register(MODULE_NAME, 'DoorControl.prototype._onMouseDown', DoorControlPrototypeOnMouseDownHandler, 'MIXED');
     //@ts-ignore
     libWrapper.register(MODULE_NAME, 'DoorControl.prototype._onRightDown', DoorControlPrototypeOnRightDownHandler, 'MIXED');
-
+  
   }
+  // else{
+  //   //@ts-ignore
+  //   libWrapper.register(MODULE_NAME, 'DoorControl.prototype._onMouseDown', DoorControlPrototypeOnMouseDownHandler2, 'WRAPPER');
+  // }
 
 }
 
@@ -107,7 +186,7 @@ export const DoorControlPrototypeOnMouseDownHandler = async function (wrapped, .
           AudioHelper.play({src: fixedPlayPath, volume: playVolume, autoplay: true, loop: false}, true);
         }
         return;
-      }
+      }           
     }
 
     // YOU NEED THIS ANYWAY FOR A STRANGE BUG WITH OVERRIDE AND SOUND OF DOOR
@@ -153,3 +232,69 @@ export const DoorControlPrototypeOnRightDownHandler = function (wrapped, ...args
   }
   return wrapped(...args);
 }
+
+// export const DoorControlPrototypeOnMouseDownHandler2 = async function (wrapped, ...args) {
+
+//   const doorControl = this;
+
+//   if(<boolean>game.settings.get(MODULE_NAME, "enableAmbientDoor")) {
+//     AmbientDoors.onDoorMouseDownCheck(doorControl);
+//   }
+
+//   return wrapped(...args);
+// }
+
+// export const DoorControlPrototypeGetTextureHandler  = async function(){//async function(wrapped, ...args) {
+
+//   const doorControl = this;
+//   let texture:PIXI.Texture;
+//   if(<boolean>game.settings.get(MODULE_NAME, "enableDesignerDoor")) {
+//     texture = await DesignerDoors.getTextureOverride(doorControl);
+//     if(texture!=null){
+//       //return texture;
+//     }else{
+//       //return wrapped(...args);
+//     }
+//   }else{
+//     //return wrapped(...args);
+//   }
+// }
+
+// export const DoorControlPrototypeDrawHandler  = async function(wrapped, ...args) {
+
+//   const doorControl = this;
+//   if(<boolean>game.settings.get(MODULE_NAME, "enableDesignerDoor")) {
+//     DesignerDoors.draw(doorControl);
+//   }
+//   return wrapped(...args);
+
+// }
+
+// export const ControlsLayerPrototypeDrawDoorsHandler = async function (wrapped, ...args) {
+//   if(<boolean>game.settings.get(MODULE_NAME, "enabledShowDoorIcons")) {
+//     ShowDoorIcons.controlsLayerPrototypeDrawDoorsHandler(this);
+//   }
+//   return wrapped(...args);
+// };
+
+// export const WallsLayerPrototypeActivateHandler = function (wrapped, ...args) {
+//   if(<boolean>game.settings.get(MODULE_NAME, "enabledShowDoorIcons")) {
+//     ShowDoorIcons.wallsLayerPrototypeActivateHandler();
+//   }
+//   return wrapped(...args);
+// }
+
+// export const WallsLayerPrototypeDeactivateHandler = function (wrapped, ...args) {
+//   if(<boolean>game.settings.get(MODULE_NAME, "enabledShowDoorIcons")) {
+//     ShowDoorIcons.wallsLayerPrototypeDeactivateHandler();
+//   }
+//   return wrapped(...args);
+// }
+
+// export const WallPrototypeOnModifyWallHandler = function (wrapped, ...args) {
+//   if(<boolean>game.settings.get(MODULE_NAME, "enabledShowDoorIcons")) {
+//     const [state] = args;
+//     ShowDoorIcons.wallPrototypeOnModifyWallHandler(state);
+//   }
+//   return wrapped(...args);
+// }